MOVEONLY: Separate CTransaction and dependencies from core
