+++ src/Makefile.am
@@ -81,6 +81,7 @@ BITCOIN_CORE_H = \
   coins.h \
   compat.h \
   core.h \
+  core/transaction.h \
   core_io.h \
   crypter.h \
   db.h \
@@ -213,6 +214,7 @@ libbitcoin_common_a_SOURCES = \
   chainparams.cpp \
   coins.cpp \
   core.cpp \
+  core/transaction.cpp \
   core_read.cpp \
   core_write.cpp \
   hash.cpp \
