+++ src/bloom.cpp
@@ -4,7 +4,7 @@
 
 #include "bloom.h"
 
-#include "core.h"
+#include "core/transaction.h"
 #include "script/script.h"
 #include "script/standard.h"
 #include "streams.h"
