+++ src/coincontrol.h
@@ -5,7 +5,7 @@
 #ifndef COINCONTROL_H
 #define COINCONTROL_H
 
-#include "core.h"
+#include "core/transaction.h"
 
 /** Coin Control Features. */
 class CCoinControl
