+++ src/coins.h
@@ -6,7 +6,8 @@
 #ifndef BITCOIN_COINS_H
 #define BITCOIN_COINS_H
 
-#include "core.h"
+#include "core.h" // Only for CTxOutCompressor
+#include "core/transaction.h"
 #include "serialize.h"
 #include "uint256.h"
