+++ src/core_read.cpp
@@ -4,7 +4,7 @@
 
 #include "core_io.h"
 
-#include "core.h"
+#include "core/transaction.h"
 #include "script/script.h"
 #include "serialize.h"
 #include "streams.h"
