+++ src/script/sign.cpp
@@ -5,7 +5,7 @@
 
 #include "script/sign.h"
 
-#include "core.h"
+#include "core/transaction.h"
 #include "key.h"
 #include "keystore.h"
 #include "script/standard.h"
