+++ src/test/main_tests.cpp
@@ -2,7 +2,7 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "core.h"
+#include "core/transaction.h"
 #include "main.h"
 
 #include <boost/test/unit_test.hpp>
