+++ src/txdb.cpp
@@ -5,7 +5,6 @@
 
 #include "txdb.h"
 
-#include "core.h"
 #include "pow.h"
 #include "uint256.h"
