+++ src/txmempool.cpp
@@ -5,7 +5,6 @@
 
 #include "txmempool.h"
 
-#include "core.h"
 #include "streams.h"
 #include "util.h"
 #include "utilmoneystr.h"
